% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.smooth.R
\name{calc.smooth}
\alias{calc.smooth}
\title{Calculate 2-d kernal smoother given input bandwidth}
\usage{
calc.smooth(x, y, data, h)
}
\arguments{
\item{x}{X location from expand.grid}

\item{y}{Y location from expand.grid}

\item{data}{Path to expand.grid dataframe}

\item{h}{Bandwidth}
}
\description{
Calculate 2-d kernal smoother given input bandwidth
}
\examples{
calc.smooth()
}
\keyword{creation}
