% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{create.plots}
\alias{create.plots}
\title{A function to export individual plots of each time-series step}
\usage{
create.plots(date, imputed.data, known.data, column, lim, path)
}
\arguments{
\item{date}{Input dataframe}

\item{imputed.data}{Path to geojson}

\item{known.data}{Input dataframe}

\item{column}{Path to geojson}

\item{lim}{Input dataframe}

\item{path}{Path to geojson}
}
\description{
A function to export individual plots of each time-series step
}
\examples{
create.plots()
}
\keyword{creation}
