% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{map.plot}
\alias{map.plot}
\title{A function to join point data to geojson}
\usage{
map.plot(date, data, column, map, title)
}
\arguments{
\item{date}{Input dataframe}

\item{data}{Path to geojson}

\item{column}{Input dataframe}

\item{map}{Path to geojson}

\item{title}{Path to geojson}
}
\description{
A function to join point data to geojson
}
\examples{
map.plot()
}
\keyword{creation}
