% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.cv.R
\name{smooth.cv}
\alias{smooth.cv}
\title{Returns the cross-validation error for an input bandwidth}
\usage{
smooth.cv(h, data)
}
\arguments{
\item{h}{Bandwidth}

\item{data}{Dataframe}
}
\description{
Returns the cross-validation error for an input bandwidth
}
\examples{
smooth.cv()
}
\keyword{creation}
