% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.join.R
\name{spatial.join}
\alias{spatial.join}
\title{A function to join point data to geojson}
\usage{
spatial.join(input, geojson)
}
\arguments{
\item{input}{Input dataframe}

\item{geojson}{Path to geojson}
}
\description{
A function to join point data to geojson
}
\examples{
spatial.join()
}
\keyword{creation}
